Move http_01.py to _internal
