+++ certbot-nginx/certbot_nginx/_internal/configurator.py
@@ -26,7 +26,7 @@ from certbot.plugins import common
 
 from certbot_nginx._internal import constants
 from certbot_nginx._internal import display_ops
-from certbot_nginx import http_01
+from certbot_nginx._internal import http_01
 from certbot_nginx import nginxparser
 from certbot_nginx import obj  # pylint: disable=unused-import
 from certbot_nginx import parser
